facer-ai/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
flask_mongoengine
flask_smorest
python_dotenv
pandas
gunicorn
cmake
pika
pyrabbit
face_recognition==1.3.0
dlib==19.22
numpy==1.23.4
opencv_python==4.6.0.66
pymongo==4.3.2